refactor call_inline_function methods
